Move bitvm related utils to bitvm_client.rs
