+++ core/src/builder/transaction/sign.rs
@@ -1,4 +1,5 @@
 use crate::actor::{Actor, WinternitzDerivationPath};
+use crate::bitvm_client::ClementineBitVMPublicKeys;
 use crate::builder;
 use crate::builder::transaction::creator::ReimburseDbCache;
 use crate::builder::transaction::{DepositData, TransactionType};
@@ -7,7 +8,6 @@ use crate::database::Database;
 use crate::errors::BridgeError;
 use crate::operator::Operator;
 use crate::rpc::clementine::{KickoffId, RawSignedTx, RawSignedTxs};
-use crate::utils::ClementineBitVMPublicKeys;
 use crate::watchtower::Watchtower;
 use bitcoin::{Transaction, XOnlyPublicKey};
