+++ core/src/test/rpc.rs
@@ -3,8 +3,8 @@
 //! This tests checks if typical RPC flows works or not.
 
 use super::common::run_single_deposit;
+use crate::bitvm_client::SECP;
 use crate::rpc::clementine::WithdrawParams;
-use crate::utils::SECP;
 use bitcoin::{secp256k1::SecretKey, Address, Amount};
 use tonic::Request;
